yatube_api/api/permissions.py
from rest_framework.permissions import IsAuthenticatedOrReadOnly, SAFE_METHODS


class IsAuthorOrReadOnly(IsAuthenticatedOrReadOnly):

    def has_object_permission(self, request, view, obj):
        if request.method in SAFE_METHODS:
            return True
        return (
            getattr(obj, 'author', None) == request.user
            or getattr(obj, 'user', None) == request.user
        )
